Move trackviewer and TrackComponent into own file
